+++ controllers/items.py
@@ -4,7 +4,7 @@ Item table
 """
 
 from models.models import Item
-from init_db import session
+import init_db
 from models.schemas import ItemSchema
 
 
@@ -15,12 +15,13 @@ def read_all():
     :return:        json string of list of items
     """
     # Create the list of items from our data
+    session = init_db.get_session()
     items = session.query(Item).order_by(Item.name).all()
 
     # Serialize the data for the response
     item_schema = ItemSchema(many=True)
     item_data = item_schema.dump(items)
-    return item_data
+    return item_data, 200
 
 
 def create(body):
@@ -30,6 +31,7 @@ def create(body):
     :param body: JSON object of the item to be created
     :return: item created if request successful, error 409 otherwise
     """
+    session = init_db.get_session()
     existing_item = (
         session.query(Item).filter(Item.name == body.get("name"))
         .one_or_none()
@@ -47,7 +49,7 @@ def create(body):
         session.commit()
 
         item_data = item_schema.dump(new_item_deserialized)
-        return item_data, 201
+        return item_data, 200
 
     # otherwise, person exists already
     else:
@@ -61,6 +63,7 @@ def read_one(item_id):
     :param item_id: id of item to find
     :return: JSON object of the item matching the id
     """
+    session = init_db.get_session()
     existing_item = (
         session.query(Item).filter(Item.id == item_id)
         .one_or_none()
@@ -69,7 +72,7 @@ def read_one(item_id):
     if existing_item is not None:
         item_schema = ItemSchema()
         item_data_serialized = item_schema.dump(existing_item)
-        return item_data_serialized
+        return item_data_serialized, 200
 
     else:
         return {"error": f"Item not found for id: {item_id}"}, 404
@@ -82,8 +85,10 @@ def update(item_id, body):
     :param body: JSON object containing new changes to the specific item
     :return: JSON object containing information about the item updated
     """
-
-    read_one(item_id)
+    session = init_db.get_session()
+    read_one_response = read_one(item_id)
+    if read_one_response[1] == 404:
+        return read_one_response
 
     # deserialize data into a database object
     item_schema = ItemSchema()
@@ -102,13 +107,16 @@ def delete(item_id):
     :param item_id: id of the item to be deleted
     :return: JSON object containing information about the item deleted
     """
+    session = init_db.get_session()
     existing_item = read_one(item_id)
+    if existing_item[1] == 404:
+        return existing_item
 
     # deserialize item to a database object
     item_schema = ItemSchema()
-    item_schema_deserialized = item_schema.load(existing_item, session=session)
+    item_schema_deserialized = item_schema.load(existing_item[0], session=session)
 
     # if the execution reaches this line, then existing item is not None
     session.delete(item_schema_deserialized)
     session.commit()
-    return existing_item
+    return existing_item[0], 200
