+++ controllers/orders.py
@@ -3,11 +3,11 @@ This is the orders module and supports all the REST actions for the
 Order table
 """
 
-from decimal import Decimal
-from controllers import customers, employees
+from sqlalchemy.exc import IntegrityError
+import init_db
+from controllers import customers, employees, payments
 from models.models import Order
-from init_db import session
-from models.schemas import OrderSchema, PaymentSchema
+from models.schemas import OrderSchema
 
 
 def read_all():
@@ -16,13 +16,14 @@ def read_all():
     with the complete lists of orders
     :return:        json string of list of orders
     """
+    session = init_db.get_session()
     # Create the list of orders from our data
     orders = session.query(Order).order_by(Order.order_time).all()
 
     # Serialize the data for the response
     order_schema = OrderSchema(many=True)
     orders_data = order_schema.dump(orders)
-    return orders_data
+    return orders_data, 200
 
 
 def create(body):
@@ -32,36 +33,25 @@ def create(body):
     :param body:  order to create
     :return:       201 on success, 406 on person exists
     """
-    if customers.read_one(body.get("customer_id")) is None:
-        return {"error": f"Order not found for Id: {body.get('customer_id')}"}, 404
+    session = init_db.get_session()
+    if customers.read_one(body.get("customer_id"))[1] == 404:
+        return {"error": f"Customer with Id {body.get('customer_id')} not found"}, 404
 
-    if employees.read_one(body.get("employee_id")) is None:
-        return {"error": f"Order not found for Id: {body.get('employee_id')}"}, 404
+    if employees.read_one(body.get("employee_id"))[1] == 404:
+        return {"error": f"Employee with Id {body.get('employee_id')} not found"}, 404
 
     order_schema = OrderSchema()
     new_order_deserialized = order_schema.load(body, session=session)
 
-    session.add(new_order_deserialized)
-    session.flush()
-
-    price = 0
-    for orderItem in new_order_deserialized.items_ordered:
-        item_details = orderItem.item_details
-        price += Decimal(item_details.price.strip('$')) * orderItem.quantity
-
-    new_payment = {
-        "order_id": new_order_deserialized.id,
-        "customer_id": body.get("customer_id"),
-        "employee_id": body.get("employee_id"),
-        "price": price
-    }
-
-    payment_schema = PaymentSchema()
-    payment_schema.load(new_payment, session=session)
-    session.commit()
+    try:
+        session.add(new_order_deserialized)
+        session.commit()
+        session.flush()
+    except IntegrityError:
+        return {"error": "order item not found for one or more inserted items"}
 
     new_order_serialize = order_schema.dump(new_order_deserialized)
-    return new_order_serialize, 201
+    return new_order_serialize, 200
 
 
 def read_one(order_id):
@@ -71,6 +61,7 @@ def read_one(order_id):
     :param order_id: id of order to find
     :return: JSON object of the order matching the id
     """
+    session = init_db.get_session()
     existing_order = (
         session.query(Order).filter(Order.id == order_id)
         .one_or_none()
@@ -79,7 +70,7 @@ def read_one(order_id):
     if existing_order is not None:
         order_schema = OrderSchema()
         order_data_serialized = order_schema.dump(existing_order)
-        return order_data_serialized
+        return order_data_serialized, 200
     else:
         return {"error": f"Order not found for Id: {order_id}"}, 404
 
@@ -91,25 +82,28 @@ def update(order_id, body):
     :param body: JSON object containing new changes to the order
     :return: JSON object of the updated order
     """
+    session = init_db.get_session()
 
-    if read_one(order_id) is None:
+    existing_order = read_one(order_id)
+
+    if existing_order[1] == 404:
         return {"error": f"Order not found for Id: {order_id}"}, 404
 
-    if customers.read_one(body.get("customer_id")) is None:
+    if customers.read_one(body.get("customer_id"))[1] == 404:
         return {"error": f"Order not found for Id: {body.get('customer_id')}"}, 404
 
-    if employees.read_one(body.get("employee_id")) is None:
+    if employees.read_one(body.get("employee_id"))[1] == 404:
         return {"error": f"Employee not found for Id: {body.get('employee_id')}"}, 404
 
     body["id"] = order_id
     order_schema = OrderSchema()
     existing_order_deserialized = order_schema.load(body, session=session)
-
     session.merge(existing_order_deserialized)
-    session.commit()
 
     new_order_serialize = order_schema.dump(existing_order_deserialized)
-    return new_order_serialize, 201
+
+    session.commit()
+    return new_order_serialize, 200
 
 
 def delete(order_id):
@@ -118,13 +112,17 @@ def delete(order_id):
     :param order_id: ID of the order to delete
     :return: JSON object of the deleted order
     """
+    session = init_db.get_session()
+
     existing_order = read_one(order_id)
+    if existing_order[1] == 404:
+        return existing_order
 
     # deserialize order to a database object
     order_schema = OrderSchema()
-    existing_order_deserialized = order_schema.load(existing_order, session=session)
+    existing_order_deserialized = order_schema.load(existing_order[0], session=session)
 
     # if the execution reaches this line, then existing order is not None
     session.delete(existing_order_deserialized)
     session.commit()
-    return existing_order
+    return existing_order[0], 200
